move util to folder
