+++ patientsearch/src/js/components/SiteLogo.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { makeStyles } from "@material-ui/core/styles";
-import { imageOK } from "./Utility";
+import { imageOK } from "../helpers/utility";
 import { useSettingContext } from "../context/SettingContextProvider";
 import theme from "../themes/theme";
