+++ patientsearch/src/js/containers/App.js
@@ -3,7 +3,7 @@ import "regenerator-runtime/runtime";
 import React, { Component } from "react";
 import Layout from "../layout/Layout";
 import PatientListTable from "../components/PatientListTable";
-import TimeoutModal from "../components/TimeoutModal.js";
+import TimeoutModal from "../components/TimeoutModal";
 import Version from "../components/Version";
 
 export default class App extends Component {
