+++ patientsearch/src/js/helpers/utility.js
